+++ packages/semver/src/builders/version/utils/git.ts
@@ -1,6 +1,9 @@
+import { BuilderContext } from '@angular-devkit/architect';
+/* @todo replace this dependency. */
+import { exec } from '@lerna/child-process';
 import * as gitRawCommits from 'git-raw-commits';
 
-import { Observable } from 'rxjs';
+import { defer, Observable, throwError } from 'rxjs';
 import { last, scan, startWith } from 'rxjs/operators';
 
 /**
@@ -27,4 +30,79 @@ export function getCommits({
     startWith([]),
     last()
   );
+}
+
+export function tryPushToGitRemote({
+  remote,
+  branch,
+  noVerify,
+  context,
+}: {
+  remote: string;
+  branch: string;
+  context: BuilderContext;
+  noVerify: boolean;
+}): Observable<void> {
+  if (remote == null || branch == null) {
+    return throwError(
+      'Missing configuration for Git push, please provide --remote and --branch options, see: https://github.com/jscutlery/semver#configure' +
+        '\n' +
+        'Skipping git push...'
+    );
+  }
+
+  return defer(() =>
+    _pushToGitRemote({
+      remote,
+      branch,
+      noVerify,
+      context,
+    })
+  );
+}
+
+export function _pushToGitRemote({
+  remote,
+  branch,
+  context,
+  noVerify,
+}: {
+  remote: string;
+  branch: string;
+  context: BuilderContext;
+  noVerify: boolean;
+}): Promise<void> {
+  const gitPushOptions = [
+    '--follow-tags',
+    ...(noVerify ? ['--no-verify'] : []),
+  ];
+
+  return exec('git', [
+    'push',
+    ...gitPushOptions,
+    '--atomic',
+    remote,
+    branch,
+  ]).catch((error) => {
+    // @see https://github.com/sindresorhus/execa/blob/v1.0.0/index.js#L159-L179
+    // the error message _should_ be on stderr except when GIT_REDIRECT_STDERR has been configured to redirect
+    // to stdout. More details in https://git-scm.com/docs/git#Documentation/git.txt-codeGITREDIRECTSTDERRcode
+    if (
+      /atomic/.test(error.stderr) ||
+      (process.env.GIT_REDIRECT_STDERR === '2>&1' &&
+        /atomic/.test(error.stdout))
+    ) {
+      // --atomic is only supported in git >=2.4.0, which some crusty CI environments deem unnecessary to upgrade.
+      // so let's try again without attempting to pass an option that is almost 5 years old as of this writing...
+      context.logger.warn('git push ' + error.stderr);
+      context.logger.info(
+        'git push --atomic failed, attempting non-atomic push'
+      );
+
+      return exec('git', ['push', ...gitPushOptions, remote, branch]);
+    }
+
+    // ensure unexpected errors still break chain
+    throw error;
+  });
 }
